% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firehose_downloader.R
\name{firehose_get}
\alias{firehose_get}
\title{Main function for downloading data from firehose.}
\usage{
firehose_get(disease_cohort, data_type, run_type = "stddata",
  run_date = "latest", resume = TRUE)
}
\arguments{
\item{disease_cohort}{use \code{list_cohorts} to check available disease cohorts.}

\item{data_type}{"mutation", "expression", or "Aggregate_AnalysisFeatures".}

\item{run_type}{use \code{list_runs} to check available run types.}

\item{run_date}{use \code{list_runs} to check valid run dates.}

\item{resume}{set TRUE to resume broken downloads.}
}
\value{
R object for firehose data.
}
\description{
Main function for downloading data from firehose.
}
\examples{
\dontrun{
mut_data <- firehose_get("HNSC", "mutation", run_date = "2015_08_21", run_type = "stddata")
exp_data <- firehose_get("HNSC", "expression", run_date = "2015_08_21", run_type = "stddata")
}
}

