% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ext.R
\docType{data}
\name{hnsc_exp_part}
\alias{hnsc_exp_part}
\title{Example data}
\format{\preformatted{'data.frame':	795 obs. of  305 variables:
 $ TCGA-BA-4074-01: num  476 32 69 10 20 ...
 $ TCGA-BA-4076-01: num  183 6 69 2 15 ...
 $ TCGA-BA-4077-01: num  153 55 196 25 94 127 21 208 102 183 ...
 $ TCGA-BA-4078-01: num  448 15 1294 3 471 ...
 $ TCGA-BA-5149-01: num  157 3 54 0 451 ...
 $ TCGA-BA-5151-01: num  74.2 17 238 0 474 15 114 2 234 10 ...
 $ TCGA-BA-5152-01: num  119 46 293 25 308 28 83 2 218 2 ...
 $ TCGA-BA-5153-01: num  154 36 1285 0 73 ...
 $ TCGA-BA-5555-01: num  82 24 455 4 337 ...
 $ TCGA-BA-5556-01: num  124 20 150 0 82 ...
 $ TCGA-BA-5557-01: num  184 13 72 0 574 38 91 76 676 40 ...
 $ TCGA-BA-5558-01: num  72 7 115 0 281 27 11 8 484 2 ...
 $ TCGA-BA-5559-01: num  265 51 2557 36 213 ...
 $ TCGA-BA-6868-01: num  37.2 31 119 11 757 ...
 $ TCGA-BA-6869-01: num  361 29 296 3 44 ...
 $ TCGA-BA-6870-01: num  186 11 160 2 109 ...
 $ TCGA-BA-6871-01: num  203 57 773 25 169 ...
 $ TCGA-BA-6872-01: num  126 5 108 0 138 ...
 $ TCGA-BA-6873-01: num  43 3 141 0 29 60 33 741 46 609 ...
 $ TCGA-BA-7269-01: num  584 21 120 3 189 80 98 2040 586 381 ...
 $ TCGA-BB-4217-01: num  197 24 473 2 7 ...
 $ TCGA-BB-4223-01: num  90 10 1050 4 39 89 24 4 631 1 ...
 $ TCGA-BB-4224-01: num  228 20 368 10 52 ...
 $ TCGA-BB-4225-01: num  78.5 30 537 0 202 ...
 $ TCGA-BB-4228-01: num  207 9 958 1 105 ...
 $ TCGA-CN-4723-01: num  219 25 554 8 582 ...
 $ TCGA-CN-4725-01: num  24 2 260 39 104 ...
 $ TCGA-CN-4726-01: num  50 6 29 0 385 12 15 914 519 127 ...
 $ TCGA-CN-4727-01: num  139 4 42 0 468 ...
 $ TCGA-CN-4728-01: num  391 25 434 6 943 ...
 $ TCGA-CN-4729-01: num  62 8 158 2 311 83 21 9 88 1 ...
 $ TCGA-CN-4730-01: num  134 9 106 3 47 6 9 15 157 6 ...
 $ TCGA-CN-4731-01: num  125 41 60 1 80 50 57 5 117 3 ...
 $ TCGA-CN-4733-01: num  102 5 396 3 139 ...
 $ TCGA-CN-4735-01: num  412 73 813 8 85 ...
 $ TCGA-CN-4736-01: num  83 14 771 44 317 ...
 $ TCGA-CN-4737-01: num  59.9 14 658 27 384 ...
 $ TCGA-CN-4738-01: num  45 116 332 23 50 105 78 10 433 5 ...
 $ TCGA-CN-4739-01: num  102 38 396 3 55 132 43 188 406 49 ...
 $ TCGA-CN-4740-01: num  129 30 660 34 475 ...
 $ TCGA-CN-4741-01: num  67 73 840 70 101 ...
 $ TCGA-CN-4742-01: num  85 11 140 3 80 ...
 $ TCGA-CN-5355-01: num  69 16 299 20 331 ...
 $ TCGA-CN-5356-01: num  462 9 554 3 473 ...
 $ TCGA-CN-5359-01: num  118 38 527 6 988 ...
 $ TCGA-CN-5360-01: num  104 149 1009 54 85 ...
 $ TCGA-CN-5363-01: num  72 0 111 0 189 ...
 $ TCGA-CN-5364-01: num  61 9 301 6 31 ...
 $ TCGA-CN-5365-01: num  155 11 42 1 62 ...
 $ TCGA-CN-5366-01: num  75.8 11 221 4 159 ...
 $ TCGA-CN-5367-01: num  66.4 17 198 10 130 ...
 $ TCGA-CN-5369-01: num  123 21 179 3 153 ...
 $ TCGA-CN-5370-01: num  68.2 138 578 32 571 ...
 $ TCGA-CN-5373-01: num  63 3 260 1 206 134 20 976 93 611 ...
 $ TCGA-CN-5374-01: num  309 13 767 42 18 ...
 $ TCGA-CN-6010-01: num  79.8 90 562 50 70 ...
 $ TCGA-CN-6011-01: num  300 1 40 2 878 ...
 $ TCGA-CN-6012-01: num  190 18 633 25 139 ...
 $ TCGA-CN-6013-01: num  74 4 253 4 21 ...
 $ TCGA-CN-6016-01: num  252 4 610 31 122 ...
 $ TCGA-CN-6018-01: num  76 11 318 5 468 ...
 $ TCGA-CN-6019-01: num  240 40 607 26 226 ...
 $ TCGA-CN-6020-01: num  77 12 97 1 84 ...
 $ TCGA-CN-6021-01: num  54.8 18 262 3 54 ...
 $ TCGA-CN-6022-01: num  209 17 546 20 120 ...
 $ TCGA-CN-6023-01: num  136 95 865 49 133 ...
 $ TCGA-CN-6024-01: num  162 32 284 8 77 ...
 $ TCGA-CN-6988-01: num  37 10 250 3 106 328 10 106 297 67 ...
 $ TCGA-CN-6989-01: num  235 35 561 11 539 ...
 $ TCGA-CN-6992-01: num  51.5 32 1119 136 280 ...
 $ TCGA-CN-6994-01: num  95.9 24 127 0 279 ...
 $ TCGA-CN-6995-01: num  46.9 5 184 1 30 ...
 $ TCGA-CN-6997-01: num  38.1 23 438 4 128 ...
 $ TCGA-CN-6998-01: num  127 5 48 18 7 ...
 $ TCGA-CQ-5323-01: num  51.9 10 410 1 119 ...
 $ TCGA-CQ-5324-01: num  30 17 41 0 127 11 56 58 86 14 ...
 $ TCGA-CQ-5325-01: num  74 3 255 5 103 ...
 $ TCGA-CQ-5326-01: num  48 10 257 0 296 25 125 55 89 0 ...
 $ TCGA-CQ-5329-01: num  245 4 170 3 53 ...
 $ TCGA-CQ-5330-01: num  44.8 1 326 10 144 ...
 $ TCGA-CQ-5331-01: num  287 6 217 2 791 ...
 $ TCGA-CQ-5332-01: num  129 19 475 10 285 33 24 11 89 2 ...
 $ TCGA-CQ-5334-01: num  57 1 103 0 402 16 12 93 724 27 ...
 $ TCGA-CQ-6218-01: num  45 15 109 1 307 51 28 22 273 3 ...
 $ TCGA-CQ-6220-01: num  91 4 109 0 112 11 26 334 18 51 ...
 $ TCGA-CQ-6221-01: num  1192 79 428 6 343 ...
 $ TCGA-CQ-6223-01: num  82 7 143 3 20 40 6 8 14 0 ...
 $ TCGA-CQ-6224-01: num  148 7 703 25 130 ...
 $ TCGA-CQ-6225-01: num  64 26 161 6 26 ...
 $ TCGA-CQ-6228-01: num  192 25 450 24 30 ...
 $ TCGA-CQ-6229-01: num  33.1 9 312 16 57 ...
 $ TCGA-CQ-7065-01: num  19 4 46 5 80 ...
 $ TCGA-CQ-7068-01: num  86.1 5 86 0 55 ...
 $ TCGA-CR-5243-01: num  95.6 4 1561 2 426 ...
 $ TCGA-CR-5247-01: num  98 6 79 0 41 ...
 $ TCGA-CR-5248-01: num  218 8 795 0 5 111 28 6 33 3 ...
 $ TCGA-CR-5250-01: num  194 6 991 0 150 ...
 $ TCGA-CR-6470-01: num  267 97 943 34 292 ...
 $ TCGA-CR-6471-01: num  63.2 3 371 3 135 ...
  [list output truncated]
}}
\usage{
hnsc_exp_part
}
\description{
Example data
}
\keyword{datasets}

