% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ext.R
\docType{data}
\name{hnsc_expressed_genes}
\alias{hnsc_expressed_genes}
\title{Example data}
\format{\preformatted{ chr [1:18863] "A1BG" "A2BP1" "A2LD1" "A2ML1" "A2M" "A4GALT" ...
}}
\usage{
hnsc_expressed_genes
}
\description{
Example data
}
\keyword{datasets}

