% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ext.R
\docType{data}
\name{hnsc_mut_part}
\alias{hnsc_mut_part}
\title{Example data}
\format{\preformatted{'data.frame':	979 obs. of  7 variables:
 $ Hugo_Symbol         : chr  "CLCA1" "CLCA1" "CLCA1" "CLCA1" ...
 $ Chromosome          : chr  "1" "1" "1" "1" ...
 $ Start_position      : int  86939133 86964330 86961284 86939149 86961316 66081724 66064395 66036415 66101886 66058502 ...
 $ End_position        : int  86939133 86964330 86961284 86939149 86961316 66081724 66064395 66036415 66101886 66058502 ...
 $ Reference_Allele    : chr  "G" "G" "G" "G" ...
 $ Tumor_Seq_Allele2   : chr  "C" "C" "A" "A" ...
 $ Tumor_Sample_Barcode: chr  "TCGA-CV-7427-01" "TCGA-CR-6481-01" "TCGA-BA-6869-01" "TCGA-D6-6516-01" ...
}}
\usage{
hnsc_mut_part
}
\description{
Example data
}
\keyword{datasets}

