% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis.R
\name{identify_expressed_genes}
\alias{identify_expressed_genes}
\title{Identify expressed genes.
Identify expressed genes based on user-provided expression data.}
\usage{
identify_expressed_genes(df_user_exp, rsem_thres = 1, frac_thres = 0.5)
}
\arguments{
\item{df_user_exp}{A dataframe.}

\item{rsem_thres}{Default value is 1.}

\item{frac_thres}{Default value is 0.5.}
}
\description{
Identify expressed genes.
Identify expressed genes based on user-provided expression data.
}
\examples{
\dontrun{
mut_data <- firehose_get("HNSC", "mutation", run_date = "2015_08_21", run_type = "stddata")
mut_data <- mut_data[[1]]
mut_sample_ids <- unique(mut_data[[7]])

exp_data <- firehose_get("HNSC", "expression", run_date = "2015_08_21", run_type = "stddata")
exp_data <- exp_data[[1]]
exp_sample_ids <- colnames(exp_data)

common_case <- intersect(mut_sample_ids, exp_sample_ids)
exp_control <- grepl("-11$", exp_sample_ids)
hnsc_exp <- exp_data[, (exp_sample_ids \%in\% common_case) | exp_control]
expressed_genes <- identify_expressed_genes(hnsc_exp, rsem_thres = 1, frac_thres = 0.5)
}
}

