% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firehose_downloader.R
\name{list_cohorts}
\alias{list_cohorts}
\title{List available cohorts on firehose website.}
\usage{
list_cohorts()
}
\value{
A character vector of cohorts.
}
\description{
List available cohorts on firehose website.
}
\examples{
\dontrun{
list_cohorts()
}
}

