% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firehose_downloader.R
\name{list_run_types}
\alias{list_run_types}
\title{List available run types on firehose website.}
\usage{
list_run_types()
}
\value{
A character vector of cohorts.
}
\description{
List available run types on firehose website.
}
\examples{
\dontrun{
list_run_types()
}
}

