% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firehose_downloader.R
\name{list_runs}
\alias{list_runs}
\title{List available runs on firehose website.}
\usage{
list_runs()
}
\value{
A character vector of runs.
}
\description{
List available runs on firehose website.
}
\examples{
\dontrun{
list_runs()
}
}

