% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_pathway.R
\name{network_from_significant_branches}
\alias{network_from_significant_branches}
\title{Generate network from pathway branches.}
\usage{
network_from_significant_branches(pathway_result, p_thres = 0.05)
}
\arguments{
\item{pathway_result}{Pathway-based analysis result.}

\item{p_thres}{P value threshold, default 0.05.}
}
\value{
A igraph object.
}
\description{
Generate network from pathway branches.
}
\examples{
\dontrun{

prepare_ma()
res.gene <- perform_gene_pathway(hnsc_mut_part, hnsc_exp_part)
res.network <- perform_network_pathway(res.gene[[2]], res.gene[[3]], hnsc_expressed_genes)

g <- network_from_significant_branches(res.network)
}
}

