% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_ppi.R
\name{perform_gene_ppi}
\alias{perform_gene_ppi}
\title{Perform gene module of PPI-based approach.}
\usage{
perform_gene_ppi(df_user_mut, df_user_exp, sid_pattern = c("-01$", "-11$"),
  raw_count = TRUE)
}
\arguments{
\item{df_user_mut}{Mutation dataframe.}

\item{df_user_exp}{Gene expression dataframe.}

\item{sid_pattern}{A regex string to match case and control IDs.}

\item{raw_count}{Logical value, set TRUE to process RSEM-based expression
data.}
}
\value{
A list.
}
\description{
Perform gene module of PPI-based approach.
}
\examples{
\dontrun{

prepare_ma()
res.gene <- perform_gene_ppi(hnsc_mut_part, hnsc_exp_part)
}
}

