% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_pathway.R
\name{perform_network_pathway}
\alias{perform_network_pathway}
\title{Perform network module of pathway-based approach.}
\usage{
perform_network_pathway(df_mut_score_wide, whole_fc_genes, expressed_genes,
  mcutoff = 2, fraction = 0.6, padj = 0.01, log2_fold_change = 2)
}
\arguments{
\item{df_mut_score_wide}{A dataframe.}

\item{whole_fc_genes}{A list.}

\item{expressed_genes}{A character vector.}

\item{mcutoff}{A float number between 0 and 6, default is 2.}

\item{fraction}{A float number between 0 and 1, default is 0.6.}

\item{padj}{Adjusted p value threshold for differential analysis, default is 0.01.}

\item{log2_fold_change}{log2 transformed fold change threshold for
differential analysis, default is 2.}
}
\value{
A list.
}
\description{
Perform network module of pathway-based approach.
}
\examples{
\dontrun{

prepare_ma()
res.gene <- perform_gene_pathway(hnsc_mut_part, hnsc_exp_part)
res.network <- perform_network_pathway(res.gene[[2]], res.gene[[3]], hnsc_expressed_genes)
}
}

