% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_ppi.R
\name{perform_network_ppi}
\alias{perform_network_ppi}
\title{Perform network module of PPI-based approach.}
\usage{
perform_network_ppi(df_mut_score_wide, whole_fc_genes, mcutoff = 2,
  fraction = 0.6, nei_order = 2, mweight = 0.6, padj = 0.01,
  log2_fold_change = 2)
}
\arguments{
\item{df_mut_score_wide}{A dataframe.}

\item{whole_fc_genes}{A list.}

\item{mcutoff}{A float number between 0 and 6, default is 2.}

\item{fraction}{A float number between 0 and 1, default is 0.6.}

\item{nei_order}{The order of neighborhood from the mutant gene, default is 2.}

\item{mweight}{weight of mutation score for combined score, default is 0.6.}

\item{padj}{Adjusted p value threshold for differential analysis, default is 0.01.}

\item{log2_fold_change}{log2 transformed fold change threshold for
differential analysis, default is 2.}
}
\value{
A list.
}
\description{
Perform network module of PPI-based approach.
}
\examples{
\dontrun{

prepare_ma()
res.gene <- perform_gene_ppi(hnsc_mut_part, hnsc_exp_part)
res.network <- perform_network_ppi(res.gene[[2]], res.gene[[3]])
}
}

