% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_pathway.R
\name{plot_pathway}
\alias{plot_pathway}
\title{visualize pathway-based network.}
\usage{
plot_pathway(net, ...)
}
\arguments{
\item{net}{An igraph object.}

\item{...}{Other parameters.}
}
\description{
Visualized pathway-based network, this function takes
\code{network_from_significant_branches} output as input.
}
\examples{
\dontrun{

prepare_ma()
res.gene <- perform_gene_pathway(hnsc_mut_part, hnsc_exp_part)
res.network <- perform_network_pathway(res.gene[[2]], res.gene[[3]], hnsc_expressed_genes)

g <- network_from_significant_branches(res.network)
plot_pathway(g)
}
}

