% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_ppi.R
\name{plot_ppi}
\alias{plot_ppi}
\title{visualize PPI-based network.}
\usage{
plot_ppi(net, ...)
}
\arguments{
\item{net}{An igraph object.}

\item{...}{Other parameters.}
}
\description{
Visualize PPI-based network, this function takes
\code{network_from_ppi} output as input.
}
\examples{
\dontrun{

prepare_ma()
res.gene <- perform_gene_ppi(hnsc_mut_part, hnsc_exp_part)
res.network <- perform_network_ppi(res.gene[[2]], res.gene[[3]])
g <- network_from_ppi(res.network)
plot_ppi(g)
}
}

