% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firehose_downloader.R
\name{prepare_ma}
\alias{prepare_ma}
\title{Prepare mutation assessor dataset.}
\usage{
prepare_ma()
}
\description{
Prepare mutation assessor dataset.
}
\examples{
\dontrun{
prepare_ma()
}
}

